+++ derive/src/derive/std/iter/trusted_len.rs
@@ -1,10 +1,8 @@
-use proc_macro2::TokenStream;
-
 use crate::utils::*;
 
 pub(crate) const NAME: &[&str] = &["TrustedLen"];
 
-pub(crate) fn derive(data: &Data) -> Result<TokenStream> {
+pub(crate) fn derive(data: &Data, stack: &mut Stack<ItemImpl>) -> Result<()> {
     derive_trait!(
         data,
         Some(ident_call_site("Item")),
@@ -13,4 +11,5 @@ pub(crate) fn derive(data: &Data) -> Result<TokenStream> {
             unsafe trait TrustedLen: ::core::iter::Iterator {}
         }?,
     )
+    .map(|item| stack.push(item))
 }